Fikset feil i datatype for telefonnummer i SQL filen som gjorde at data ikke ble lagret i tabellen

+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE Billett
     id INTEGER AUTO_INCREMENT NOT NULL,
     fornavn VARCHAR(255) NOT NULL,
     etternavn VARCHAR(255) NOT NULL,
-    telefonnr SMALLINT NOT NULL,
+    telefonnr INTEGER NOT NULL,
     epost VARCHAR(255) NOT NULL,
     film  VARCHAR(255) NOT NULL,
     antall SMALLINT NOT NULL,
